ptrace_disable: no need to disable stepping
ptrace_disable() is called when the tracee is quiescent and we
are going to untrace. This means we are going to clear ->ptrace
and wake up the tracee. Now that ptrace_stop() checks PT_ bits
and does user_disable_single_step() we can remove this code from
ptrace_disable(), it is unneeded and not utrace-friendly.

TODO: change !x86 code.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>